[MIG-CO-REQ] Explicitly cast utime_t to double in output for simpler usage

+++ src/mds/MDCache.cc
@@ -12030,7 +12030,7 @@ C_MDS_RetryRequest::C_MDS_RetryRequest(MDCache *c, MDRequestRef& r)
 {
   if (r->fromServer) {
     MClientRequest *req = mdr->client_request;
-    dout(0) << ADSLTAG_MIGRATION_CORRE_REQUEST << " " << ceph_clock_now() << " Pending request: " << ceph_mds_op_name(req->get_op()) << " on " << req->get_path() << dendl;
+    dout(0) << ADSLTAG_MIGRATION_CORRE_REQUEST << " " << (double)ceph_clock_now() << " Pending request: " << ceph_mds_op_name(req->get_op()) << " on " << req->get_path() << dendl;
   }
 }
 #else
